refactor(gui): import kernel through barrel
