+++ packages/gui/src/main/local-composition-root.ts
@@ -1,8 +1,8 @@
 import path from "node:path";
 import { makeLocalLifecycleEngine } from "../../../adapters/local/src/index.ts";
 import { makeLocalControllerService } from "../../../application/src/index.ts";
-import type { HarnessLayoutOverrides } from "../../../kernel/src/layout/index.ts";
-import { createHarnessRuntimeContext, resolveHarnessRuntimeContext } from "../../../kernel/src/layout/index.ts";
+import type { HarnessLayoutOverrides } from "../../../kernel/src/index.ts";
+import { createHarnessRuntimeContext, resolveHarnessRuntimeContext } from "../../../kernel/src/index.ts";
 import { createGuiServiceBridgeForService } from "../api/service-bridge.ts";
 import type { GuiServiceBridge } from "../api/service-bridge.ts";
